move src/quick-query to src/local-queries/quick-query
